# m8flow-backend/src/m8flow_backend/api.yml
openapi: "3.0.2"
info:
  version: 1.0.0
  title: m8flow-backend-extension
paths:
...
  /process-models/{modified_process_model_identifier}/files/{file_name}:
    get:
      summary: Process Modeler — read one file's raw content
      description: >-
        Raw bytes of a single named file (.bpmn/.dmn/etc.) inside a process
        model directory, for opening it in the designer's modeler. Concrete
        tenant required. Denied or missing model/file -> 404.
      operationId: m8flow_backend.routes.processes_controller.get_process_model_file
      tags:
        - Processes
      parameters:
        - name: modified_process_model_identifier
          in: path
          required: true
          schema:
            type: string
          description: "Process model path with / replaced by :"
        - name: file_name
          in: path
          required: true
          schema:
            type: string
        - name: tenantId
          in: query
          schema:
            type: string
          description: Super-admin-only concrete tenant override.
      responses:
        "200":
          description: Raw file content
          content:
            application/octet-stream: {}
        "400":
          description: Missing concrete tenant or invalid file name
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized - user not authenticated
        "404":
          description: Process model or file not found, or not permitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      summary: Process Modeler — save one file's content
      description: >-
        Writes one named file's bytes into an existing process model
        directory. .bpmn content is re-imported (same behavior as
        POST /process-models); every file type is git-committed
        (best-effort). Concrete tenant required. Denied -> 403; missing
        model -> 404.
      operationId: m8flow_backend.routes.processes_controller.put_process_model_file
      tags:
        - Processes
      parameters:
        - name: modified_process_model_identifier
          in: path
          required: true
          schema:
            type: string
          description: "Process model path with / replaced by :"
        - name: file_name
          in: path
          required: true
          schema:
            type: string
        - name: tenantId
          in: query
          schema:
            type: string
          description: Super-admin-only concrete tenant override.
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: File saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessModelFileSaveResponse"
        "400":
          description: >-
            Missing concrete tenant, invalid file name, empty body, or
            unsupported BPMN construct
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized - user not authenticated
        "403":
          description: Not permitted to modify this process model
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Process model not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Process-model overview — delete a file
      description: >-
        Deletes one named file. The primary file is 409 until another file is
        set as primary. Super-admin is always 403. No git.
      operationId: m8flow_backend.routes.processes_controller.delete_process_model_file
      tags:
        - Processes
      parameters:
        - name: modified_process_model_identifier
          in: path
          required: true
          schema:
            type: string
          description: "Process model path with / replaced by :"
        - name: file_name
          in: path
          required: true
          schema:
            type: string
        - name: tenantId
          in: query
          schema:
            type: string
          description: Super-admin-only concrete tenant override (writes still 403).
      responses:
        "200":
          description: File deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessModelFileDeleteResponse"
        "400":
          description: Invalid or reserved file name
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized - user not authenticated
        "403":
          description: Not permitted to modify this process model
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Process model or file not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: File is the primary file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /templates/{id}/files/{file_name}:
    get:
      summary: Get a template file by name
      description: Download a single file (BPMN, JSON, DMN, etc.) by filename
      operationId: m8flow_backend.routes.templates_controller.template_get_file
      tags:
        - Template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: file_name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File content
          content:
            application/xml:
              schema:
                type: string
            application/json:
              schema:
                type: string
            text/markdown:
              schema:
                type: string
        "404":
          description: Template or file not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      summary: Update a template file by name
      description: Update content of a single file. Template must not be published.
      operationId: m8flow_backend.routes.templates_controller.template_put_file
      tags:
        - Template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: file_name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: string
          application/xml:
            schema:
              type: string
          text/markdown:
            schema:
              type: string
          text/plain:
            schema:
              type: string
      responses:
        "200":
          description: File updated, returns the updated template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
        "400":
          description: Bad request
        "403":
          description: Template is published
        "404":
          description: Template or file not found
        "500":
          description: Internal server error
    delete:
      summary: Delete a template file by name
      description: Remove a file from the template. Template must not be published. Cannot delete last file or only BPMN.
      operationId: m8flow_backend.routes.templates_controller.template_delete_file
      tags:
        - Template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: file_name
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: File deleted
        "403":
          description: Forbidden (published template or last file or only BPMN)
        "404":
          description: Template or file not found
        "500":
          description: Internal server error
...
  /connector-templates/{connector_type}:
    parameters:
      - name: connector_type
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: m8flow_backend.routes.connector_profiles_controller.connector_template_show
      summary: One connector schema
      tags:
        - Connectors
      responses:
        "200":
          description: Connector descriptor
          content:
            application/json:
              schema:
                type: object
        "404":
          description: Unknown connector type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /connector-profiles:
    get:
      operationId: m8flow_backend.routes.connector_profiles_controller.connector_profile_list
      summary: Connector profiles for the active tenant
      tags:
        - Connectors
      parameters:
        - name: connector_type
          in: query
          required: false
          schema:
            type: string
        - name: include_inactive
          in: query
          required: false
          schema:
            type: boolean
            default: true
      responses:
        "200":
          description: Profiles. Secret values are never included.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
    post:
      operationId: m8flow_backend.routes.connector_profiles_controller.connector_profile_create
      summary: Create a connector profile
      tags:
        - Connectors
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          description: The created profile, references only
          content:
            application/json:
              schema:
                type: object
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: A profile of that name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /connector-profiles/{profile_id}:
    parameters:
      - name: profile_id
        in: path
        required: true
        schema:
          type: integer
    get:
      operationId: m8flow_backend.routes.connector_profiles_controller.connector_profile_show
      summary: One connector profile
      tags:
        - Connectors
      responses:
        "200":
          description: The profile, references only
          content:
            application/json:
              schema:
                type: object
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: m8flow_backend.routes.connector_profiles_controller.connector_profile_update
      summary: Update a connector profile
      tags:
        - Connectors
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: The updated profile
          content:
            application/json:
              schema:
                type: object
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: m8flow_backend.routes.connector_profiles_controller.connector_profile_delete
      summary: Deactivate a connector profile, or remove it with hard=true
      tags:
        - Connectors
      parameters:
        - name: hard
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Deactivated, or removed
          content:
            application/json:
              schema:
                type: object
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ProcessModelFileSaveResponse:
      type: object
      required:
        - name
        - size_bytes
        - updated_at_in_seconds
      properties:
        name:
          type: string
        size_bytes:
          type: integer
        updated_at_in_seconds:
          type: integer
...
    ProcessModelFileDeleteResponse:
      type: object
      required:
        - deleted
        - name
      properties:
        deleted:
          type: boolean
        name:
          type: string
...
    Template:
      type: object
      properties:
        id:
          type: integer
        templateKey:
          type: string
        version:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        category:
          type: string
        tenantId:
          type: string
        visibility:
          type: string
          enum: [PRIVATE, TENANT, PUBLIC]
        files:
          type: array
          items:
            type: object
            properties:
              fileType:
                type: string
                enum: [bpmn, json, dmn, md]
              fileName:
                type: string
        isPublished:
          type: boolean
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
        updatedAt:
          type: string
          format: date-time
        modifiedBy:
          type: string
        bpmnContent:
          type: string
        is_deleted:
          type: boolean
...
    ErrorData:
      type: object
      properties:
        error_code:
          type: string
        message:
          type: string
    ErrorResponse:
      $ref: "#/components/schemas/ErrorData"
...